+++ certificate/letsencrypt/dns/artfiles/provider.go
@@ -1,0 +1,57 @@
+package artfiles
+
+import (
+	"context"
+
+	"github.com/go-acme/lego/v4/challenge"
+	"github.com/go-acme/lego/v4/providers/dns/artfiles"
+
+	"dillmann.com.br/nginx-ignition/certificate/letsencrypt/dns"
+	"dillmann.com.br/nginx-ignition/core/common/dynamicfields"
+	"dillmann.com.br/nginx-ignition/core/common/i18n"
+)
+
+const (
+	usernameFieldID = "artfilesUsername"
+	passwordFieldID = "artfilesPassword"
+)
+
+type Provider struct{}
+
+func (p *Provider) ID() string {
+	return "ARTFILES"
+}
+
+func (p *Provider) Name(ctx context.Context) *i18n.Message {
+	return i18n.M(ctx, i18n.K.CertificateLetsencryptDnsArtfilesName)
+}
+
+func (p *Provider) DynamicFields(ctx context.Context) []dynamicfields.DynamicField {
+	return dns.LinkedToProvider(p.ID(), []dynamicfields.DynamicField{
+		{
+			ID:          usernameFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsArtfilesUsername),
+			Required:    true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+		{
+			ID:          passwordFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsArtfilesPassword),
+			Required:    true,
+			Sensitive:   true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+	})
+}
+
+func (p *Provider) ChallengeProvider(
+	_ context.Context,
+	_ []string,
+	parameters map[string]any,
+) (challenge.Provider, error) {
+	config := artfiles.NewDefaultConfig()
+	config.Username, _ = parameters[usernameFieldID].(string)
+	config.Password, _ = parameters[passwordFieldID].(string)
+
+	return artfiles.NewDNSProviderConfig(config)
+}