+++ certificate/letsencrypt/dns/czechia/provider.go
@@ -1,0 +1,49 @@
+package czechia
+
+import (
+	"context"
+
+	"github.com/go-acme/lego/v4/challenge"
+	"github.com/go-acme/lego/v4/providers/dns/czechia"
+
+	"dillmann.com.br/nginx-ignition/certificate/letsencrypt/dns"
+	"dillmann.com.br/nginx-ignition/core/common/dynamicfields"
+	"dillmann.com.br/nginx-ignition/core/common/i18n"
+)
+
+const (
+	tokenFieldID = "czechiaToken"
+)
+
+type Provider struct{}
+
+func (p *Provider) ID() string {
+	return "CZECHIA"
+}
+
+func (p *Provider) Name(ctx context.Context) *i18n.Message {
+	return i18n.M(ctx, i18n.K.CertificateLetsencryptDnsCzechiaName)
+}
+
+func (p *Provider) DynamicFields(ctx context.Context) []dynamicfields.DynamicField {
+	return dns.LinkedToProvider(p.ID(), []dynamicfields.DynamicField{
+		{
+			ID:          tokenFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsCzechiaToken),
+			Required:    true,
+			Sensitive:   true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+	})
+}
+
+func (p *Provider) ChallengeProvider(
+	_ context.Context,
+	_ []string,
+	parameters map[string]any,
+) (challenge.Provider, error) {
+	config := czechia.NewDefaultConfig()
+	config.Token, _ = parameters[tokenFieldID].(string)
+
+	return czechia.NewDNSProviderConfig(config)
+}