+++ certificate/letsencrypt/dns/leaseweb/provider.go
@@ -1,0 +1,49 @@
+package leaseweb
+
+import (
+	"context"
+
+	"github.com/go-acme/lego/v4/challenge"
+	"github.com/go-acme/lego/v4/providers/dns/leaseweb"
+
+	"dillmann.com.br/nginx-ignition/certificate/letsencrypt/dns"
+	"dillmann.com.br/nginx-ignition/core/common/dynamicfields"
+	"dillmann.com.br/nginx-ignition/core/common/i18n"
+)
+
+const (
+	apiKeyFieldID = "leasewebApiKey" // nolint:gosec
+)
+
+type Provider struct{}
+
+func (p *Provider) ID() string {
+	return "LEASEWEB"
+}
+
+func (p *Provider) Name(ctx context.Context) *i18n.Message {
+	return i18n.M(ctx, i18n.K.CertificateLetsencryptDnsLeasewebName)
+}
+
+func (p *Provider) DynamicFields(ctx context.Context) []dynamicfields.DynamicField {
+	return dns.LinkedToProvider(p.ID(), []dynamicfields.DynamicField{
+		{
+			ID:          apiKeyFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsLeasewebApiKey),
+			Required:    true,
+			Sensitive:   true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+	})
+}
+
+func (p *Provider) ChallengeProvider(
+	_ context.Context,
+	_ []string,
+	parameters map[string]any,
+) (challenge.Provider, error) {
+	config := leaseweb.NewDefaultConfig()
+	config.APIKey, _ = parameters[apiKeyFieldID].(string)
+
+	return leaseweb.NewDNSProviderConfig(config)
+}